Dependency requirements, settings, and api additions for IPDetails
