+++ REQUIREMENTS.txt
@@ -1,0 +1,3 @@
+Django==1.8.4
+requests==2.6.2
+djangorestframework==3.1.1
