+++ api_test/api/urls.py
@@ -3,6 +3,6 @@ from rest_framework import routers
 from .views import IPDetailsView, Traffic
 
 urlpatterns = patterns('api_test.api.views',
-    url(r'/threat/ip/(?P<ip>.*)$', IPDetailsView.as_view(), name='api_threat_details'),
-    url(r'/traffic$', Traffic.as_view(), namespace='api_traffic')
+    url(r'threat/ip/(?P<ip>.*)$', IPDetailsView.as_view(), name='threat_details'),
+    url(r'traffic$', Traffic.as_view(), name='traffic')
 )
