+++ api_test/settings.py
@@ -37,7 +37,8 @@ INSTALLED_APPS = (
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'api_test.api'
+    'api_test.api',
+    'api_test.datastore'
 )
 
 MIDDLEWARE_CLASSES = (
