+++ api_test/urls.py
@@ -1,9 +1,9 @@
 from django.conf.urls import url, include
 from rest_framework import routers
-from server import APIRoot
+from .views import APIRoot
 
 urlpatterns = [
     url(r'^$', APIRoot.as_view(), name='root'),
 
-    url(r'^api/', include('api.urls', namespace="api"))
+    url(r'^api/', include('api_test.api.urls', namespace="api"))
 ]
